% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytokdynamics-package.R
\docType{package}
\name{cytokdynamics-package}
\alias{cytokdynamics}
\alias{cytokdynamics-package}
\title{cytokdynamics: What the Package Does (Title Case)}
\description{
This package will be able to analyse gene expression data in plants, concentrating on Cytokinin dependent negative feedback behavior.
}
\keyword{internal}
