% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_negative_feedback.R
\name{is_negative_feedback}
\alias{is_negative_feedback}
\title{is_negative_feedback Feedback identification}
\usage{
is_negative_feedback(
  data,
  relative_quantity = "relative_quantity_ratio",
  time = "time",
  group = "group",
  max_time = 120,
  filter_args = list()
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{relative_quantity}{The name of the column containing the relative quantity data.}

\item{time}{The name of the column containing the time data.}

\item{group}{The name of the column containing the group data.}

\item{max_time}{The maximum time to consider in the analysis. If NULL, all data is used. By default it is set to <= 120 min.}

\item{filter_args}{A list of additional arguments to filter the data. You could for example filter out specific groups or time points. Take in mind, that if group is not specified or group column has less than 2 values, nls will be used instead of nlme.}
}
\value{
An nlme object representing the fitted model.
}
\description{
This function uses a specified nlme to detect whether a system exhibits negative feedback behavior. The specified nlme follows an oscillatory curve, starting with a peak (A), followed by an oscillatoric behavior. The formula is as follows: relative_quantity ~ A * sin(omega * time + phase).
Take in mind: if the output is an error, this most likely indicates that we dont have a negative feedback dynamic. Check out cytokdynamics vignette to see how to define if there are other dynamic behaviors.
}
\examples{
# use the package intern datasets cytokinin expression:
data(cytokinin_expression)
is_negative_feedback(cytokinin_expression)
is_negative_feedback(data = cytokinin_expression, filter_args = list('group == "500_direct_ARR5"'))
}
